Last update

+++ src/main/java/com/example/plantsstore/controllers/ProductController.java
@@ -4,27 +4,17 @@ import java.util.ArrayList;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.*;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.*;
 
-import com.example.plantsstore.DTO.requests.ProductRequestDTO;
-import com.example.plantsstore.DTO.responses.ProductResponseDTO;
-import com.example.plantsstore.DTO.responses.SupplierResponseDTO;
-import com.example.plantsstore.model.Product;
-import com.example.plantsstore.model.Supplier;
-import com.example.plantsstore.repository.ProductRepository;
-import com.example.plantsstore.repository.SupplierRepository;
-import com.example.plantsstore.utility.Utility;
+import com.example.plantsstore.DTO.requests.*;
+import com.example.plantsstore.DTO.responses.*;
+import com.example.plantsstore.model.*;
+import com.example.plantsstore.repository.*;
+import com.example.plantsstore.services.*;
+import com.example.plantsstore.utility.*;
 
 @Controller
 @RequestMapping(path = "/products")
@@ -33,93 +23,70 @@ public class ProductController {
   private ProductRepository productRepository;
   @Autowired
   private SupplierRepository supplierRepository;
+  @Autowired
+  private SupplyRequestItemRepository supplyRequestItemRepository;
 
   @PostMapping("/saveProduct")
-  public @ResponseBody ResponseEntity<?> saveProduct(@RequestBody ProductRequestDTO.Product newProduct) {
-    Product existingProdcut = !Utility.isStringEmpty(newProduct.id)
-      ? productRepository.findByNameAndIdNot(newProduct.name, newProduct.id)
-      : productRepository.findByName(newProduct.name);
-
+  public @ResponseBody ResponseEntity<?> saveProduct(@RequestBody ProductRequest.NewProduct newProduct) {
     try {
-      if (existingProdcut != null) {
-        throw new Exception("Such product already exists!");
-      }
+      Product product = !Utility.isStringEmpty(newProduct.getId())
+          ? productRepository.getById(newProduct.getId())
+          : new Product();
 
-      Product product = !Utility.isStringEmpty(newProduct.id) ? productRepository.getById(newProduct.id) : null;
+      product.setName(newProduct.getName());
+      product.setDescription(newProduct.getDescription());
+      product.setImage(newProduct.getImage());
 
-      if (product != null) {
-        product.setName(newProduct.name);
-        product.setImage(newProduct.image);
-        product.setDescription(newProduct.description);
-        product.setPrice(newProduct.price);
-        product.setSuppliers(supplierRepository.findAllByIdIn(newProduct.suppliers));
-      } else {
-        product = new Product(
-          newProduct.name,
-          newProduct.description,
-          newProduct.price,
-          newProduct.image,
-          supplierRepository.findAllByIdIn(newProduct.suppliers)
-        );
+      List<Supplier> suppliers = new ArrayList<>();
+      if (!newProduct.getSuppliers().isEmpty()) {
+        suppliers = supplierRepository.findAllByIdIn(newProduct.getSuppliers());
       }
+      product.setSuppliers(suppliers);
 
       productRepository.save(product);
+
+      return ResponseEntity.status(200).body("Success");
     } catch (Exception ex) {
       return ResponseEntity.status(400).body(ex.getMessage());
     }
-
-    return ResponseEntity.status(200).body("Success");
   }
 
   @GetMapping("/getProducts")
   public @ResponseBody ResponseEntity<?> getProducts(
-    @RequestParam("limit") Integer limit,
-    @RequestParam("offset") Integer offset,
-    @RequestParam(value = "query", required = false) String query
-  ) {
-    ProductResponseDTO.Products result = new ProductResponseDTO.Products();
+      @RequestParam(value = "limit") Integer limit,
+      @RequestParam(value = "offset") Integer offset,
+      @RequestParam(value = "query", required = false) String query) {
     try {
+      ProductResponse.ProductsPagination result = new ProductResponse.ProductsPagination();
+
       Pageable pageable = PageRequest.of(offset, limit);
       Page<Product> productsPage = Utility.isStringEmpty(query)
-        ? productRepository.findAllByOrderByCreatedAtDesc(pageable)
-        : productRepository.findByNameContainingOrderByCreatedAtDesc(query, pageable);
+          ? productRepository.findAllByOrderByCreatedAtDesc(pageable)
+          : productRepository.findByNameContainingOrderByCreatedAtDesc(query, pageable);
 
       List<Product> productContent = productsPage.getContent();
-      List<ProductResponseDTO.Product> products = new ArrayList<>();
+
+      List<ProductResponse.Product> products = new ArrayList<>();
       productContent.forEach(product -> {
-        products.add(new ProductResponseDTO.Product(
-          product.getId(),
-          product.getName(),
-          product.getDescription(),
-          product.getPrice(),
-          product.getImage(),
-          processSuppliers(product.getSuppliers())
-        ));
+        ProductResponse.Product productResponse = new ProductResponse.Product();
+        productResponse.setId(product.getId());
+        productResponse.setName(product.getName());
+        productResponse.setDescription(product.getDescription());
+        productResponse.setImage(product.getImage());
+        productResponse.setPrice(product.getPrice());
+        productResponse.setSuppliers(ProductService.getProductSuppliers(product.getSuppliers()));
+
+        products.add(productResponse);
       });
 
-      result = new ProductResponseDTO.Products(
-        (int) productsPage.getTotalElements(),
-        products
-      );
+      result = new ProductResponse.ProductsPagination(
+          (int) productsPage.getTotalElements(),
+          products);
+
+      return ResponseEntity.status(200).body(result);
     } catch (Exception ex) {
       return ResponseEntity.status(400).body(ex.getMessage());
     }
-    return ResponseEntity.status(200).body(result);
-  }
-
-  private static List<SupplierResponseDTO.Supplier> processSuppliers(List<Supplier> suppliers) {
-    List<SupplierResponseDTO.Supplier> result = new ArrayList<>();
-    suppliers.forEach(supplier -> {
-      result.add(new SupplierResponseDTO.Supplier(
-        supplier.getId(),
-        supplier.getEmail(),
-        supplier.getName(),
-        supplier.getPhone(),
-        supplier.getContactPerson()
-      ));
-    });
-
-    return result;
   }
 
   @DeleteMapping("/deleteProduct")
@@ -133,6 +100,11 @@ public class ProductController {
     }
 
     try {
+      List<SupplyRequestItem> supplyRequestItems = supplyRequestItemRepository.getAllByProductId(productId);
+      if (!supplyRequestItems.isEmpty()) {
+        return ResponseEntity.status(400).body("Can not remove product due to it connected with request.");
+      }
+
       productRepository.deleteById(productId);
     } catch (Exception ex) {
       return ResponseEntity.status(400).body(ex.getMessage());
